stomp: inline default handlers and move server/client assembly
